tests/integration/requirements.txt
aio-pika==9.4.1
aiormq==6.8.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fastapi==0.110.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
motor==3.4.0
multidict==6.0.5
pamqp==3.3.0
pika==1.3.2
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
pytest==8.2.0
python-dotenv==1.0.1
pytest-asyncio==0.23.6